Deleted the hidden layer
Deleted the hidden layer and separated the main.py file to make code readable
