+++ tutorial2-code/AbstractCar.py
@@ -1,0 +1,237 @@
+import numpy
+import pygame
+import numpy as np
+import math
+import pandas as pd
+from utils import *
+
+
+TRACK_LINES = [((35, 130), (35, 650)), ((35, 650), (160, 770)), ((160, 770), (670, 770)), ((670, 770), (760, 640)),
+               ((760, 640), (760, 110)), ((760, 110), (760, 20)), ((760, 20), (370, 20)), ((170, 200), (170, 600)),
+               ((170, 600), (240, 660)), ((240, 660), (580, 670)), ((580, 670), (660, 590)), ((660, 590), (660, 190)),
+               ((35, 130), (35, 20)), ((35, 20), (300, 20)), ((300, 20), (300, 110)), ((300, 110), (300, 510)),
+               ((370, 20), (370, 90)), ((370, 90), (370, 240)), ((370, 240), (420, 240)), ((420, 240), (500, 240)),
+               ((660, 190), (660, 130)), ((660, 130), (500, 130)), ((500, 240), (560, 240)), ((560, 240), (560, 510)),
+               ((560, 510), (500, 560)), ((500, 560), (330, 560)), ((330, 560), (300, 510))]
+
+BONUS_LINES = [((173, 216), (296, 219)), ((171, 201), (299, 112)), ((170, 200), (175, 20)), ((39, 131), (168, 202)),
+               ((40, 250), (167, 250)), ((167, 326), (38, 321)), ((37, 427), (165, 427)), ((167, 558), (37, 559)),
+               ((190, 621), (100, 708)), ((249, 662), (252, 769)), ((344, 666), (347, 764)), ((420, 667), (434, 764)),
+               ((535, 673), (548, 767)), ((582, 672), (671, 768)), ((627, 626), (717, 692)), ((660, 590), (760, 641)),
+               ((662, 502), (762, 501)), ((662, 402), (754, 405)), ((660, 313), (763, 306)), ((661, 224), (759, 218)),
+               ((662, 189), (758, 106)), ((760, 20), (660, 130)), ((621, 130), (669, 25)), ((588, 22), (583, 126)),
+               ((498, 129), (488, 25)), ((374, 91), (501, 132)), ((420, 242), (498, 133)), ((510, 260), (660, 130)),
+               ((561, 304), (657, 286)), ((566, 369), (660, 366)), ((561, 439), (658, 444)), ((563, 510), (658, 523)),
+               ((534, 541), (620, 623)), ((464, 567), (465, 660)), ((354, 564), (356, 659)), ((316, 545), (212, 627)),
+               ((174, 503), (297, 496)), ((294, 414), (170, 405))]
+
+RED_CAR = scale_image(pygame.image.load("imgs/red-car.png"), 0.45)
+
+WIN = pygame.display.set_mode((810, 810))
+
+pygame.display.set_caption("Racing Game!")
+
+FPS = 60
+
+INPUT_LAYER_SHAPE = (6, 1)
+
+WI_SHAPE = (4, 6)
+BI_SHAPE = (1, 4)
+
+W1_SHAPE = (4, 4)
+B1_SHAPE = (1, 4)
+
+
+class AbstractCar:
+    def __init__(self, _max_vel, _rotation_vel, _wi=None, _bi=None, _w1=None, _b1=None):
+        self.img = self.IMG
+        self.max_vel = _max_vel
+        self.vel = 0
+        self.rotation_vel = _rotation_vel
+        self.angle = 0
+        self.x, self.y = self.START_POS
+        self.acceleration = 0.1
+        self.score = 0
+        self.index_of_bonus_line = 0
+        self.next_bonus_line = BONUS_LINES[self.index_of_bonus_line]
+        self.sensors = []
+        self.input_layer = numpy.zeros(INPUT_LAYER_SHAPE)
+        self.points_sensor = []
+        self.rotation_side = 0  # 0 is right 1 is left
+        if _wi is not None:
+            self.weights_input_layer = _wi
+        else:
+            self.weights_input_layer = (2 * np.random.rand(*WI_SHAPE) - 1)
+
+        if _bi is not None:
+            self.bias_input_layer = _bi
+        else:
+            self.bias_input_layer = (2 * np.random.rand(*BI_SHAPE) - 1)
+        if _w1 is not None:
+            self.weights_l1 = _w1
+        else:
+            self.weights_l1 = (2 * np.random.rand(*W1_SHAPE) - 1)
+        if _b1 is not None:
+            self.bias_l1 = _b1
+        else:
+            self.bias_l1 = (2 * np.random.rand(*B1_SHAPE) - 1)
+
+    def rotate(self, _left=False, _right=False):
+        if _left:
+            self.angle += self.rotation_vel
+        elif _right:
+            self.angle -= self.rotation_vel
+
+    def draw(self, _win):
+        blit_rotate_center(_win, self.img, (self.x, self.y), self.angle)
+        self.sensors = draw_sensors(_win, self.img, (self.x, self.y), self.angle)
+
+    def move_forward(self):
+        self.vel = min(self.vel + self.acceleration, self.max_vel)
+        self.move()
+
+    def move_backward(self):
+        self.vel = max(self.vel - self.acceleration, -self.max_vel)
+        self.move()
+
+    def move(self):
+        radians = math.radians(self.angle)
+        vertical = math.cos(radians) * self.vel
+        horizontal = math.sin(radians) * self.vel
+
+        self.y -= vertical
+        self.x -= horizontal
+
+    def save_model(self):
+        with open('model.csv', 'a') as csvfile:
+            np.savetxt(csvfile, self.weights_input_layer, delimiter=',', header="Input Layer Weights")
+            np.savetxt(csvfile, self.bias_input_layer, delimiter=',', header='Input Layer Bias')
+            np.savetxt(csvfile, self.weights_l1, delimiter=',', header='Layer 1 Weights')
+            np.savetxt(csvfile, self.bias_l1, delimiter=',', header='Layer 1 Bias')
+
+    def print_model(self):
+        print("Wi:", self.weights_input_layer)
+        print("Bi:", self.bias_input_layer)
+        print("W1:", self.weights_l1)
+        print("B1:", self.bias_l1)
+
+    def collide(self):
+        car_rect = self.img.get_rect(x=self.x, y=self.y)
+        for line in TRACK_LINES:
+            if car_rect.clipline(line):
+                return True
+        return False
+
+    def sense(self):
+        points = []
+        distances = []
+        self.sensors = [self.sensors[0], self.sensors[1], self.sensors[2], self.sensors[-1], self.sensors[-2]]
+        for sensor in self.sensors:
+            sensed_point = None
+            shortest_point = None
+            for track_line in TRACK_LINES:
+
+                (x1, y1), (x2, y2) = track_line
+                (x3, y3), (x4, y4) = sensor
+                x1 = int(round(x1))
+                x2 = int(round(x2))
+                x3 = int(round(x3))
+                x4 = int(round(x4))
+                y1 = int(round(y1))
+                y2 = int(round(y2))
+                y3 = int(round(y3))
+                y4 = int(round(y4))
+                if (max(x1, x2) < min(x3, x4)) or (min(x1, x2) > max(x3, x4)) \
+                        or (max(y1, y2) < min(y3, y4)) or (min(y1, y2) > max(y3, y4)):
+                    pass
+                else:
+                    if x1 == x2 and x3 == x4:
+                        pass
+
+                    elif x1 == x2:
+
+                        y = int(((y4 - y3) / (x4 - x3)) * (x1 - x3) + y3)
+                        if min(y1, y2) <= y <= max(y1, y2):
+                            dist = math.sqrt((x1 - x3) ** 2 + (y - y3) ** 2)
+                            if shortest_point is None or shortest_point > dist:
+                                shortest_point = dist
+                                sensed_point = ((0, 0, 255), (x1, y), 5)
+                    elif x3 == x4:
+                        y = int(((y2 - y1) / (x2 - x1)) * (x3 - x1) + y1)
+                        if min(y3, y4) <= y <= max(y3, y4):
+                            dist = math.sqrt((x3 - x3) ** 2 + (y3 - y) ** 2)
+                            if shortest_point is None or shortest_point > dist:
+                                shortest_point = dist
+                                sensed_point = ((255, 0, 0), (x3, y), 5)
+                    else:
+                        b1 = ((y2 - y1) / (x2 - x1))
+                        b2 = ((y4 - y3) / (x4 - x3))
+                        if b1 != b2:
+                            x = ((b2 * x3 - b1 * x1 + y1 - y3) / (b2 - b1))
+                            y = (y1 + b1 * (x - x1))
+                            if min(y3, y4) <= y <= max(y3, y4) and min(x3, x4) <= x <= max(x3, x4) \
+                                    and min(x1, x2) <= x <= max(x1, x2):
+                                dist = math.sqrt((x - x3) ** 2 + (y - y3) ** 2)
+                                if shortest_point is None or shortest_point > dist:
+                                    shortest_point = dist
+                                    sensed_point = ((0, 255, 0), (x, y), 5)
+            if sensed_point is not None and shortest_point is not None:
+                points.append(sensed_point)
+                distances.append(shortest_point)
+            else:
+                points.append(None)
+                distances.append(400)
+        self.input_layer = np.array(distances)
+        self.points_sensor = points
+
+    def reset(self):
+        self.x, self.y = self.START_POS
+        self.angle = 0
+        self.vel = 0
+        self.index_of_bonus_line = 0
+        self.next_bonus_line = BONUS_LINES[0]
+        self.score = 0
+
+    def take_action(self):
+        weighted_sum_input_layer = np.dot(self.weights_input_layer, self.input_layer) + self.bias_input_layer
+        output_layer = relu(weighted_sum_input_layer)
+        # weighted_sum_layer_1 = np.dot(self.weights_l1, input_layer_results.T)
+        # weighted_sum_layer_1 += self.bias_l1.T
+
+        # output_layer = relu(weighted_sum_layer_1)
+
+        decided_action = np.argmax(output_layer)
+
+        if decided_action == 0:
+            self.rotate(_left=True)
+            # self.move_forward()
+
+        elif decided_action == 1:
+            self.rotate(_right=True)
+            # self.move_forward()
+
+        elif decided_action == 2:
+            self.move_forward()
+
+        elif decided_action == 3:
+            self.move_backward()
+
+    def update_score(self, _score):
+        self.score += _score
+
+
+class PlayerCar(AbstractCar):
+    IMG = RED_CAR
+    START_POS = (237, 311)
+
+    def reduce_speed(self):
+        self.vel = max(self.vel - self.acceleration / 2, 0)
+        self.move()
+
+    def bounce(self):
+        self.vel = -self.vel
+        self.move()
+
+    def crash(self):
+        self.score -= 500
+
